ghostty-vt: fix infinite loop in RefCountedSet rehash threshold
RefCountedSet::add compared living < cap*0.9 in floating point; upstream
(ref_counted_set.zig) truncates the threshold to an integer via
@intFromFloat. At small capacities the difference is fatal: the default
hyperlink set (cap 3) with 2 living items reported NeedsRehash, but a
same-capacity rehash clone reclaims nothing when every allocated ID is
living, so Screen::start_hyperlink retried forever — the third unique
implicit hyperlink on a page hung. With truncation a fully-living set
always reports OutOfMemory, which grows hyperlink_bytes and makes
progress. Styles were never affected (cap >= 16 keeps cap-1 above the
truncated threshold), but manual_style_update now also routes
NeedsRehash to the same-capacity rehash clone instead of doubling the
style capacity, matching upstream manualStyleUpdate.

Regression tests: three unique implicit hyperlinks (the hang repro), and
a style-set rehash that compacts dead IDs without growing capacity.
Restores print_slice_differential_fuzz_vs_print op counts to upstream
500/500/500/200 (held down solely by this bug).

Validated: cargo test -p ghostty-vt (default and --features
slow_runtime_safety), cargo check --workspace, cargo fmt --check,
cargo clippy --workspace.

+++ crates/ghostty-vt/src/page/ref_set.rs
@@ -311,10 +311,15 @@ where
                 return Err(AddError::OutOfMemory);
             }
 
-            // Need a fresh ID.
+            // Need a fresh ID. The threshold is truncated to an integer,
+            // matching upstream's `@intFromFloat` — this matters at small
+            // capacities: when every allocated ID is living, a rehash
+            // reclaims nothing, so we must report OutOfMemory (grow) instead
+            // of NeedsRehash or the caller's rehash-retry loop never
+            // terminates (e.g. cap 3 with 2 living: 2 < trunc(2.7) is false).
             if self.next_id.to_usize() >= self.layout.cap {
                 let rehash_threshold = 0.9;
-                if (self.living as f64) < self.layout.cap as f64 * rehash_threshold {
+                if self.living < (self.layout.cap as f64 * rehash_threshold) as usize {
                     return Err(AddError::NeedsRehash);
                 }
                 return Err(AddError::OutOfMemory);
